Extract common functions from lwaftr check
