+++ src/program/lwaftr/check/check.lua
@@ -1,26 +1,22 @@
 module(..., package.seeall)
 
-local config = require("core.config")
-local counter = require("core.counter")
 local lib = require("core.lib")
-local lwconf = require('apps.lwaftr.conf')
-local lwcounter = require("apps.lwaftr.lwcounter")
-local lwutil = require("apps.lwaftr.lwutil")
+local lwconf = require("apps.lwaftr.conf")
 local setup = require("program.lwaftr.setup")
+local util = require("program.lwaftr.check.util")
 
--- Get the counter directory and names from the code, so that any change
--- in there will be automatically picked up by the tests.
-local counter_names = lwcounter.counter_names
-local counters_dir = lwcounter.counters_dir
+local load_requested_counters = util.load_requested_counters
+local read_counters = util.read_counters
+local diff_counters = util.diff_counters
+local validate_diff = util.validate_diff
+local regen_counters = util.regen_counters
 
-local write_to_file = lwutil.write_to_file
-
-function show_usage(code)
+local function show_usage(code)
    print(require("program.lwaftr.check.README_inc"))
    main.exit(code)
 end
 
-function parse_args (args)
+local function parse_args (args)
    local handlers = {}
    local opts = {}
    function handlers.h() show_usage(0) end
@@ -36,66 +32,6 @@ function parse_args (args)
    if #args ~= 5 and #args ~= 6 then show_usage(1) end
    if not opts["duration"] then opts["duration"] = 0.10 end
    return opts, args
-end
-
-function load_requested_counters(counters)
-   local result = dofile(counters)
-   assert(type(result) == "table", "Not a valid counters file: "..counters)
-   return result
-end
-
-function read_counters(c)
-   local results = {}
-   for _, name in ipairs(counter_names) do
-      local cnt = counter.open(counters_dir .. name .. ".counter", "readonly")
-      results[name] = counter.read(cnt)
-   end
-   return results
-end
-
-function diff_counters(final, initial)
-   local results = {}
-   for name, ref in pairs(initial) do
-      local cur = final[name]
-      if cur ~= ref then
-         results[name] = tonumber(cur - ref)
-      end
-   end
-   return results
-end
-
-function validate_diff(actual, expected)
-   if not lib.equal(actual, expected) then
-      local msg
-      print('--- Expected (actual values in brackets, if any)')
-      for k, v in pairs(expected) do
-         msg = k..' = '..v
-         if actual[k] ~= nil then
-            msg = msg..' ('..actual[k]..')'
-         end
-         print(msg)
-      end
-      print('--- actual (expected values in brackets, if any)')
-      for k, v in pairs(actual) do
-         msg = k..' = '..v
-         if expected[k] ~= nil then
-            msg = msg..' ('..expected[k]..')'
-         end
-         print(msg)
-      end
-      error('counters did not match')
-   end
-end
-
-local function regen_counters(counters, outfile)
-   local cnames = lwutil.keys(counters)
-   table.sort(cnames)
-   local out_val = {'return {'}
-   for _,k in ipairs(cnames) do
-      table.insert(out_val, string.format('   ["%s"] = %s,', k, counters[k]))
-   end
-   table.insert(out_val, '}\n')
-   write_to_file(outfile, (table.concat(out_val, '\n')))
 end
 
 function run(args)
